requirements.txt
django
gunicorn
psycopg2
django-heroku
requests
python-dotenv
whitenoise
dj-database-url
djangorestframework